+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+### Added
+- module instance support
+
 ## [1.0.0] - 2019-11-27
 ### Added
 - Initial version